SocialDistribution/requirements.txt
asgiref==3.5.0
beautifulsoup4==4.10.0
bleach==4.1.0
dj-database-url==0.5.0
Django==3.1.6
django-markdownify==0.9.0
django-on-heroku==1.1.2
django-rest-framework==0.1.0
djangorestframework==3.13.1
gunicorn==20.1.0
importlib-metadata==4.11.3
Markdown==3.3.6
packaging==21.3
Pillow==9.0.1
psycopg2-binary==2.9.3
pyparsing==3.0.7
pytz==2021.3
six==1.16.0
soupsieve==2.3.1
sqlparse==0.4.2
webencodings==0.5.1
whitenoise==6.0.0
yapf==0.32.0
zipp==3.7.0
